+++ addon_service/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ApiConfig(AppConfig):
+class AddonServiceConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "addon_service"
