+++ addon_service/internal_user/models.py
@@ -3,12 +3,12 @@ from django.db import models
 from addon_service.utils.base_model import AddonsServiceBaseModel
 
 class InternalUser(AddonsServiceBaseModel):
-    user_osf_uri = models.URLField(unique=True, db_index=True, null=False)
+    user_uri = models.URLField(unique=True, db_index=True, null=False)
 
     def __str__(self):
         return self.user_guid
 
     class Meta:
-        verbose_name = "OSF User"
-        verbose_name_plural = "OSF Users"
+        verbose_name = "Internal User"
+        verbose_name_plural = "Internal Users"
         app_label = "addon_service"
